# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VetHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /users/register-for-vet:
    post:
      operationId: registerForVet
      tags: [users]
      summary: Register current user as vet owner (creates a clinic)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClinicInput"
      responses:
        "201":
          description: Created clinic and updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Clinic"

  # ── Clinics ───────────────────────────────────────────────────────────────
...
  /pets:
    get:
      operationId: listMyPets
      tags: [pets]
      summary: List all pets for the current pet owner
      responses:
        "200":
          description: Pets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pet"
    post:
      operationId: createPet
      tags: [pets]
      summary: Add a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetInput"
      responses:
        "201":
          description: Created pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
  /pets/{petId}:
    get:
      operationId: getPet
      tags: [pets]
      summary: Get a pet by ID
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Pet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
    patch:
      operationId: updatePet
      tags: [pets]
      summary: Update pet details
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetUpdate"
      responses:
        "200":
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
  /pets/{petId}/status:
    patch:
      operationId: updatePetStatus
      tags: [pets]
      summary: Change pet status
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetStatusUpdate"
      responses:
        "200":
          description: Updated pet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"

  # ── Monitoring ────────────────────────────────────────────────────────────
  /pets/{petId}/monitoring:
    get:
      operationId: listMonitoring
      tags: [monitoring]
      summary: Get latest 10 monitoring records for a pet
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Monitoring records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonitoringRecord"
    post:
      operationId: addMonitoring
      tags: [monitoring]
      summary: Add a monitoring record
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoringInput"
      responses:
        "201":
          description: Created monitoring record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitoringRecord"

  # ── Vet: Search pet owners ────────────────────────────────────────────────
...
  /pets/{petId}/vaccinations:
    get:
      operationId: listVaccinations
      tags: [vaccinations]
      summary: List all vaccinations for a pet
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Vaccination records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vaccination"
    post:
      operationId: addVaccination
      tags: [vaccinations]
      summary: Add a vaccination record (vet only, date can be backdated)
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaccinationInput"
      responses:
        "201":
          description: Created vaccination record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccination"
...
  /pets/{petId}/visits:
    get:
      operationId: listVisits
      tags: [visits]
      summary: List all visits for a pet
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Visits list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visit"
    post:
      operationId: createVisit
      tags: [visits]
      summary: Create a new hospital visit
      parameters:
        - name: petId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitInput"
      responses:
        "201":
          description: Created visit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
...
  /products/{productId}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── Reports / Analytics ───────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: integer }
        stackId: { type: string }
        name: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        role:
          type: string
          enum: [pet_owner, vet, vet_owner, both]
        isPetOwner: { type: boolean }
        isVet: { type: boolean }
        isVetOwner: { type: boolean }
        clinicId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, stackId, role, isPetOwner, isVet, isVetOwner, createdAt]
    UserProfileUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
    Clinic:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        ownerId: { type: integer }
        createdAt: { type: string }
      required: [id, name, ownerId, createdAt]
    ClinicInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
      required: [name]
...
    Pet:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        dateOfBirth: { type: ["string", "null"] }
        gender:
          type: string
          enum: [male, female, unknown]
        sterilized: { type: boolean }
        color: { type: ["string", "null"] }
        speciesId: { type: integer }
        speciesName: { type: ["string", "null"] }
        ownerId: { type: integer }
        ownerName: { type: ["string", "null"] }
        ownerPhone: { type: ["string", "null"] }
        status:
          type: string
          enum: [healthy, sick, hospitalized, need_intensive_care, passed_away]
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string }
      required:
        [id, name, gender, sterilized, speciesId, ownerId, status, createdAt]
    PetInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        dateOfBirth: { type: string }
        gender:
          type: string
          enum: [male, female, unknown]
        sterilized: { type: boolean }
        color: { type: string }
        speciesId: { type: integer }
        photoUrl: { type: string }
      required: [name, gender, sterilized, speciesId]
    PetUpdate:
      type: object
      properties:
        name: { type: string }
        dateOfBirth: { type: string }
        gender:
          type: string
          enum: [male, female, unknown]
        sterilized: { type: boolean }
        color: { type: string }
        speciesId: { type: integer }
        photoUrl: { type: string }
    PetStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, sick, hospitalized, need_intensive_care, passed_away]
      required: [status]
...
    MonitoringRecord:
      type: object
      properties:
        id: { type: integer }
        petId: { type: integer }
        weight: { type: ["number", "null"] }
        height: { type: ["number", "null"] }
        temperature: { type: ["number", "null"] }
        recordedAt: { type: string }
        recordedBy: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [id, petId, recordedAt]
    MonitoringInput:
      type: object
      properties:
        weight: { type: number }
        height: { type: number }
        temperature: { type: number }
        notes: { type: string }
      required: []
    Visit:
      type: object
      properties:
        id: { type: integer }
        petId: { type: integer }
        petName: { type: ["string", "null"] }
        clinicId: { type: integer }
        vetId: { type: ["integer", "null"] }
        vetName: { type: ["string", "null"] }
        type:
          type: string
          enum: [inpatient, outpatient]
        status:
          type: string
          enum: [active, completed, cancelled]
        anamnesis: { type: ["string", "null"] }
        therapy: { type: ["string", "null"] }
        visitDate: { type: string }
        dischargeDate: { type: ["string", "null"] }
        deposit: { type: ["number", "null"] }
        totalCost: { type: number }
        createdAt: { type: string }
      required:
        [id, petId, clinicId, type, status, visitDate, totalCost, createdAt]
...
    VisitInput:
      type: object
      properties:
        clinicId: { type: integer }
        vetId: { type: integer }
        type:
          type: string
          enum: [inpatient, outpatient]
        anamnesis: { type: string }
        therapy: { type: string }
        visitDate: { type: string }
      required: [clinicId, type]
...
    Product:
      type: object
      properties:
        id: { type: integer }
        clinicId: { type: integer }
        name: { type: string }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        price: { type: number }
        stock: { type: ["integer", "null"] }
        unit: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
      required: [id, clinicId, name, price, isActive, createdAt]
...
    ProductUpdate:
      type: object
      properties:
        name: { type: string }
        category: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
        unit: { type: string }
        isActive: { type: boolean }
    Vaccination:
      type: object
      properties:
        id: { type: integer }
        petId: { type: integer }
        vaccineName: { type: string }
        brand: { type: ["string", "null"] }
        date: { type: string, description: "Date administered YYYY-MM-DD" }
        nextDueDate: { type: ["string", "null"] }
        batchNumber: { type: ["string", "null"] }
        administeredBy: { type: ["string", "null"] }
        cost: { type: ["number", "null"] }
        notes: { type: ["string", "null"] }
        vetId: { type: ["integer", "null"] }
        createdAt: { type: string }
      required: [id, petId, vaccineName, date, createdAt]
    VaccinationInput:
      type: object
      properties:
        vaccineName: { type: string, minLength: 1 }
        brand: { type: string }
        date:
          {
            type: string,
            description: "Date administered YYYY-MM-DD (can be backdated)",
          }
        nextDueDate: { type: string }
        batchNumber: { type: string }
        administeredBy: { type: string }
        cost: { type: number }
        notes: { type: string }
      required: [vaccineName, date]
...